Consolidate duplicate utilities across management commands
- Add scopes param to get_authorized_youtube_client() and use it in
  backfill_youtube_sociallinks, removing duplicate auth implementation
- Replace date_str() in create_monthly_playlist with shared parse_month()
- Replace inline month-end calculation in list_monthly_performers with
  get_month_end()
- Extract format_duration/format_schedule_time/format_schedule_price to
  houses/formatting.py, replacing duplicates in list commands
